cleanup: app-target slop-review pass
Six small behavior-preserving items from the app-target slop review,
plus one observable change (better dark-mode pill contrast).

- Drop AppDelegate.currentCameraDisplay (dead @Published with no readers).
- Drop DevicePortability.isLikelyPortable + portableKeywords (dead
  code; only the test file referenced them. Live app code uses
  portabilityCategory / importantCategory). Removes the second
  drift-prone keyword list.
- Move VersionInfo from SettingsView.swift into AboutView.swift, where
  its only consumer lives.
- Rewrite SettingsStore.profileSwitchCount doc — it powers the
  first-class "Auto-switches" Stats row, not an "easter-egg stats line".
- Drop Theme.Color.muted. CLAUDE.md says status colors are the only
  Theme.Color entries; muted = .gray isn't a status. The one consumer
  uses Color.gray inline via the new StatusPill.
- Extract StatusPill view and migrate the three inline pill chromes
  (AddProfileView's local pill helper + SettingsView "Active" + "Debug
  override"). The wizard pills lose their hardcoded .black foreground
  and 0.85 background opacity; settling on white-on-tint matches the
  two settings sites and reads cleanly in dark mode.

Net: -127 / +52 (incl. new StatusPill.swift). Test count: 184 → 180
(four dead isLikelyPortable tests removed).
