+++ Sources/AVPainRelieverApp/Theme.swift
@@ -8,9 +8,8 @@ import AppKit
 /// direction (2026-05-02): the app should look like a plain native
 /// macOS utility, no custom accent colors. So this namespace exposes
 /// a small set of semantic system colors — success / warn / error
-/// for status pills and banner-style errors, plus muted for
-/// low-volume informational pills. Everything else (headers,
-/// captions, links, button tints) uses SwiftUI's defaults —
+/// for status pills and banner-style errors. Everything else
+/// (headers, captions, links, button tints) uses SwiftUI's defaults —
 /// `.primary` for body text, `.secondary` for hints, the system
 /// accent for prominent buttons.
 ///
@@ -29,13 +28,6 @@ enum Theme {
         /// icon. System red, same as Apple's own destructive action
         /// styling.
         static let error = SwiftUI.Color.red
-        /// Quiet, informational pills — "Travels with you" tag on
-        /// portable peripherals in the wizard. Lower visual volume
-        /// than the warn / success / error tints so the
-        /// informational signal doesn't compete with the actionable
-        /// pills (green Important, yellow error banners). System
-        /// gray adapts to light/dark mode automatically.
-        static let muted = SwiftUI.Color.gray
     }
 
     enum Symbol {
